Implement allocator initialization method
The init method receives as input the available system memory. It then
calculates the required space (in bytes) for storing the bitmap and
begins a scan to all available memory blocks looking for a block large
enough to hold the bitmap data.

Once that block is found, the bitmap slice data pointers are overlaid to
that address and the memory contents are set to 0xFF (all bitmap bits set to 1).

Finally, the allocator visits all free regions that can fit into a
MaxPageOrder block and explicitly marks the appropriate MaxPageOrder
bitmap locations as available while at the same time increasing the
MaxPageOrder free count.