Use runtime.g0, runtime.m0 and replace hardcoded offsets
The rt0 code implements a dedicated function for initializing the Go runtime
structures. Instead of reserving space for a dummy g struct, the rt0
code now uses the g0 and m0 symbols defined by the runtime package. In
addition to setting up g0, the rt0 also sets up the m0 struct and links
it to g0.

Setting up m0 is a requirement for properly bootstapping the
malloc-related code in the following commits